Refactor camera control to use Cinemachine for improved functionality
