+++ MechaMorph/Assets/EnemyAi/Prefabs/BossEnemy.prefab
@@ -21,7 +21,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 1
+  m_IsActive: 0
 --- !u!4 &3557107582409442695
 Transform:
   m_ObjectHideFlags: 0
