+++ MechaMorph/Assets/EnemyAi/Prefabs/Bullet.prefab
@@ -19,7 +19,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 1
+  m_IsActive: 0
 --- !u!4 &1846644250883177177
 Transform:
   m_ObjectHideFlags: 0
