--- MechaMorph/Assets/Scripts/Ability/DashAbility.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d30e011e12902bb49860c4dd37609507
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
