+++ MechaMorph/Packages/manifest.json
@@ -1,6 +1,7 @@
 {
   "dependencies": {
     "com.unity.ai.navigation": "1.1.5",
+    "com.unity.cinemachine": "2.9.7",
     "com.unity.collab-proxy": "2.6.0",
     "com.unity.feature.development": "1.0.1",
     "com.unity.postprocessing": "3.2.2",
